+++ core/blockchain_insert.go
@@ -25,13 +25,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 )
 
-var (
-	totalGasUsed, totalGasUsedPre = 0.0, 0.0
-	totalElapsed, totalElapsedPre = 0.0, 0.0
-	totalBlocks, totalBlocksPre   = 0, 0
-	PreCommitFlag                 = false
-)
-
 // insertStats tracks and reports on block insertion.
 type insertStats struct {
 	queued, processed, ignored int
@@ -61,33 +54,11 @@ func (st *insertStats) report(chain []*types.Block, index int, dirty common.Stor
 		}
 		end := chain[index]
 
-		if PreCommitFlag {
-			totalGasUsedPre += float64(st.usedGas)
-			totalElapsedPre += float64(elapsed)
-			totalBlocksPre += st.processed
-		} else {
-			totalGasUsed += float64(st.usedGas)
-			totalElapsed += float64(elapsed)
-			totalBlocks += st.processed
-		}
-
 		// Assemble the log context and send it to the logger
 		context := []interface{}{
 			"blocks", st.processed, "txs", txs, "mgas", float64(st.usedGas) / 1000000,
 			"elapsed", common.PrettyDuration(elapsed), "mgasps", float64(st.usedGas) * 1000 / float64(elapsed),
 			"number", end.Number(), "hash", end.Hash(),
-		}
-		if PreCommitFlag {
-			totalGasUsedPre += float64(st.usedGas)
-			totalElapsedPre += float64(elapsed)
-			totalBlocksPre += st.processed
-			context = append(context, []interface{}{"t_mgasps_pre", totalGasUsedPre * 1000 / totalElapsedPre, "t_blocks_pre", totalBlocksPre}...)
-
-		} else {
-			totalGasUsed += float64(st.usedGas)
-			totalElapsed += float64(elapsed)
-			totalBlocks += st.processed
-			context = append(context, []interface{}{"t_mgasps", totalGasUsed * 1000 / totalElapsed, "t_blocks", totalBlocks}...)
 		}
 
 		if timestamp := time.Unix(int64(end.Time()), 0); time.Since(timestamp) > time.Minute {
